% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBeforFileExtension.R
\name{addBeforeFileExtension}
\alias{addBeforeFileExtension}
\title{Add Text Before File-Extension}
\usage{
addBeforeFileExtension(
  x,
  add,
  sep = "_",
  silent = FALSE,
  callFrom = NULL,
  debug = FALSE
)
}
\arguments{
\item{x}{main character vector of file-names}

\item{add}{character vector to be added}

\item{sep}{(character) separator between 'x' & 'add' (character, length 1)}

\item{silent}{(logical) suppress messages}

\item{callFrom}{(character) allow easier tracking of messages produced}

\item{debug}{(logical) additional messages for debugging}
}
\value{
modified character vector
}
\description{
This function helps changing character strings like file-names and allows adding the character vector 'add'
(length 1) before the extension (defined by last '.') of the input string 'x'. 
Used for easily creating variants/additional filenames but keeping current extension.
}
\details{
Since this function replaces the deprecated \code{addBeforFileExtension} (to correct a typo in the name of the old version of the function).
}
\examples{
addBeforeFileExtension(c("abd.txt","ghg.ijij.txt","kjh"),"new")
}
