% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjBy2ptReg.R
\name{adjBy2ptReg}
\alias{adjBy2ptReg}
\title{Adjust Values By Two-Point Regression}
\usage{
adjBy2ptReg(
  dat,
  lims,
  regrTo = c(0.1, 0.9),
  refLines = NULL,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{dat}{A numeric vector or matrix of values to adjust.}

\item{lims}{A list of two numeric vectors, each of length 1 or 2, representing the lower and upper limits for rescaling.
Example: `lims = list(c(5, 9), c(60, 90))`.}

\item{regrTo}{A numeric vector of length 2 defining the target range for rescaling (default: `c(0.1, 0.9)`).}

\item{refLines}{Optional numeric or character vector specifying reference lines (rows) in `dat` to use for rescaling.
If `NULL`, all rows are used.}

\item{silent}{Logical. If `TRUE`, suppresses messages (default: `FALSE`).}

\item{debug}{Logical. If `TRUE`, prints debugging messages (default: `FALSE`).}

\item{callFrom}{Character string for tracking messages (default: `NULL`).}
}
\value{
A numeric vector or matrix with adjusted values.
}
\description{
This function performs linear rescaling of numeric data based on specified limits (ie a specified window).
Values are adjusted to fit within a target range defined by `regrTo`.
}
\details{
Since this function shares sigificant overlap with \code{\link{scaleXY}} (which has more advanced options) it is rather suggested to use _scaleXY()_ instead. 
The function _adjBy2ptReg()_ will may deprecated in the future.
}
\examples{
set.seed(2016); dat1 <- round(runif(50, 0, 100), 1)
adjBy2ptReg(dat1, lims = list(c(5, 9), c(60, 90)))
plot(dat1, adjBy2ptReg(dat1, lims=list(c(5,9), c(60,90))))

}
