% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confInt.R
\name{confInt}
\alias{confInt}
\title{Confidence Interval To Given Alpha}
\usage{
confInt(x, alpha = 0.05, distrib = "Normal", silent = FALSE, callFrom = NULL)
}
\arguments{
\item{x}{(numeric) main input}

\item{alpha}{(numeric) significance level, accepted type I error}

\item{distrib}{(character) distribution, so far only \code{Normal} is implemented}

\item{silent}{(logical) suppress messages}

\item{callFrom}{(character) allow easier tracking of message(s) produced}
}
\value{
This function returns the confidence interval to a given \code{alpha} under the hypothesis of the Normal distribution.
}
\description{
This little function returns the confidence interval associated to a given significance level \code{alpha} under the hypothesis of the Normal distribution is valid.
}
\examples{

confInt(c(5,2:6))

}
\seealso{
\code{\link[stats]{TDist}}; \code{\link[stats]{confint}}
}
