% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffCombin.R
\name{diffCombin}
\alias{diffCombin}
\title{Compute Matrix Of Differences For All Pairwise Combinations Of Numeric Vector}
\usage{
diffCombin(
  x,
  diagAsNA = FALSE,
  prefix = TRUE,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{x}{numeric vector to compute differences for all combinations}

\item{diagAsNA}{(logical) return all self-self combinations as NA (otherwise 0)}

\item{prefix}{(logical) if TRUE, dimnames of output will specify orientation (prefix='from.' and 'to.')}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allows easier tracking of messages produced}
}
\value{
This function returns a numeric matrix of all pairwise differences
}
\description{
\code{diffCombin} returns matrix of differences (eg resulting from subsititution) for all pairwise combinations of numeric vector 'x'.
}
\examples{
diffCombin(c(10,11.1,13.3,16.6))
}
\seealso{
\code{\link[base]{diff}} for simple differences
}
