% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutAtMultSites.R
\name{.addLetterWoLast}
\alias{.addLetterWoLast}
\title{Add letter to all elements but not last}
\usage{
.addLetterWoLast(x, addChr)
}
\arguments{
\item{x}{(character) main input}

\item{addChr}{(character)}
}
\value{
This function returns a modified character vector
}
\description{
This function allows to add 'addChr' to all entries, without the last entry
}
\examples{
.addLetterWoLast(c("abc","efgh"),"Z")
}
\seealso{
\code{\link[base]{paste}}; used in \code{\link{cutAtMultSites}}
}
