% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMAList.R
\name{.allRatioMatr1to2}
\alias{.allRatioMatr1to2}
\title{Calculate ratios for each column to each column of reference-matrix}
\usage{
.allRatioMatr1to2(x, y, asLog2 = TRUE, sumMeth = "mean", callFrom = NULL)
}
\arguments{
\item{x}{(matrix or data.frame) main input1}

\item{y}{(matrix or data.frame) main input2}

\item{asLog2}{(logical)}

\item{sumMeth}{(character) method}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a numeric vector or matrix in dimension of 'x' (so far summarize all ratios from mult division from mult ref cols as mean or median )
}
\description{
This function calculates ratio(s) for each column of matrix 'x' versus all/each column(s) of matrix 'y' (reference)
}
\examples{
.allRatioMatr1to2(matrix(11:14, ncol=2), matrix(21:24, ncol=2))
}
\seealso{
\code{\link{makeMAList}}, \code{\link[base]{grep}}
}
