% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMAList.R
\name{.allRatios}
\alias{.allRatios}
\title{Search character-string and cut either before or after}
\usage{
.allRatios(dat, ty = "log2", colNaSep = "_")
}
\arguments{
\item{dat}{(matrix or data.frame) main input}

\item{ty}{(character) type of ratio (eg 'log2')}

\item{colNaSep}{(character) separator}
}
\value{
This function returns a numeric vector
}
\description{
This function extracts/cuts text-fragments out of \code{txt} following specific anchors defined by arguments \code{cutFrom} and \code{cutTo}.
}
\examples{
.allRatios(matrix(11:14, ncol=2))
}
\seealso{
\code{\link{makeMAList}}, \code{\link[base]{grep}}
}
