% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutArrayInCluLike.R
\name{.arrLstMean}
\alias{.arrLstMean}
\title{Summarize along columns of multiple arrays in list}
\usage{
.arrLstMean(
  arrLst,
  sumType = "mean",
  arrOutp = FALSE,
  signifDig = 3,
  formatCheck = FALSE,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{arrLst}{(list) main input}

\item{sumType}{(character)}

\item{arrOutp}{(logical)}

\item{signifDig}{(integer)}

\item{formatCheck}{(logical)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
array (1st dim will be summary along cols, rows will be layers of 3rd array-dim
}
\description{
This function allows summarizing along columns of multiple arrays in list
}
\examples{
.datSlope(c(3:6))
}
\seealso{
used in \code{\link{cutArrayInCluLike}}
}
