% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutArrayInCluLike.R
\name{.arrLstSEM}
\alias{.arrLstSEM}
\title{Summarize along columns of mult arrays in list}
\usage{
.arrLstSEM(
  arrLst,
  arrOutp = FALSE,
  signifDig = 3,
  formatCheck = FALSE,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{arrLst}{(list) main input}

\item{arrOutp}{(logical)}

\item{signifDig}{(integer)}

\item{formatCheck}{(logical)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
array (1st dim will be summary along cols, rows will be layers of 3rd array-dim ie dim(arrLst[[1]])[3])
}
\description{
This function allows summarizing along columns of mult arrays in list
}
\examples{
.datSlope(c(3:6))
}
\seealso{
used in \code{\link{cutArrayInCluLike}}
}
