% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asSepList.R
\name{.asDF2}
\alias{.asDF2}
\title{Convert anything to data.frame}
\usage{
.asDF2(z)
}
\arguments{
\item{z}{(numeric vector, factor, matrix or list) main input}
}
\value{
data.frame
}
\description{
This function allows converting anything to data.frame
}
\examples{
.asDF2(c(3:6))
}
\seealso{
\code{\link[base]{as.data.frame}}
}
