% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioAllComb.R
\name{.breakInSer}
\alias{.breakInSer}
\title{Get series of values after last discontinuity}
\usage{
.breakInSer(x, getFrom = "last")
}
\arguments{
\item{x}{(numeric) main input}

\item{getFrom}{(character)}
}
\value{
This function returns a numeric vector of reduced length
}
\description{
This function aims to get series of values after last discontinuity
}
\examples{
.breakInSer(c(11:14,16:18))
}
\seealso{
\code{\link[stats]{dist}}
}
