% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVectLength.R
\name{.checkArgNa}
\alias{.checkArgNa}
\title{Check argument names}
\usage{
.checkArgNa(x, argNa, lazyEval = TRUE)
}
\arguments{
\item{x}{(character) main input}

\item{argNa}{(character) argument name}

\item{lazyEval}{(logical) decide if argument should be avaluated with abbreviated names, too}
}
\value{
This function returns a elongated character vector
}
\description{
This function allows checking of argument names
}
\examples{
.checkArgNa("Abc",c("ab","Ab","BCD"))
}
\seealso{
\code{\link[base]{chartr}}
}
