% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutArrayInCluLike.R
\name{.checkConsistentArrList}
\alias{.checkConsistentArrList}
\title{Check list of arrays for consistent dimensions of all arrays}
\usage{
.checkConsistentArrList(
  arrLst,
  arrNDim = 3,
  fxName = NULL,
  varName = NULL,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{arrLst}{(list) main input}

\item{arrNDim}{(integer) number of dimensions for arrays}

\item{fxName}{(character) this name will be given in message}

\item{varName}{(character)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
list
}
\description{
This function allows to check list of arrays for consistent dimensions of all arrays
}
\examples{
.datSlope(c(3:6))
}
\seealso{
used in \code{\link{cutArrayInCluLike}}
}
