% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equLenNumber.R
\name{.checkConvt2Vect}
\alias{.checkConvt2Vect}
\title{Convert to simple vector (similar to unlist)}
\usage{
.checkConvt2Vect(dat, toNumeric = TRUE)
}
\arguments{
\item{dat}{(list, data.frame) main input}

\item{toNumeric}{(logical)}
}
\value{
character (or numeric) vector
}
\description{
This function allows converting 'dat' (may be list, data.frame etc) to simple vector, more elaborate than unlist()
}
\examples{
aa <- matrix(11:14, ncol=2)
.checkConvt2Vect(aa)
}
\seealso{
\code{\link[base]{unlist}}; used in  \code{\link{equLenNumber}}
}
