% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVectLength.R
\name{.checkFactor}
\alias{.checkFactor}
\title{Check Factor}
\usage{
.checkFactor(
  fac,
  facNa = NULL,
  minLev = 2,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{fac}{(factor) main input}

\item{facNa}{(character) level-names}

\item{minLev}{(integer) minium number of levels}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a corrceted/adjusted factor
}
\description{
This function was designed to check a factor object
}
\examples{
.checkFactor(gl(3,2))
}
\seealso{
\code{\link[base]{factor}}
}
