% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioAllComb.R
\name{.checkLegendLoc}
\alias{.checkLegendLoc}
\title{Check argument for Location of legend}
\usage{
.checkLegendLoc(
  legLoc,
  defLoc = "topright",
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{legLoc}{(character) main input}

\item{defLoc}{(character)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a character vector designing the potential location of legend
}
\description{
This function allows checking an argument for Location of legend,
if value provided not found as valid, it returns 'defLoc
}
\examples{
.checkLegendLoc("abc")
}
\seealso{
\code{\link[graphics]{legend}}
}
