% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchLinesAtGivenSlope.R
\name{.checkLmConfInt}
\alias{.checkLmConfInt}
\title{Compare 'dat' to confindence interval of linare model 'lMod' (eg from lm())}
\usage{
.checkLmConfInt(dat, lMod, level = 0.95)
}
\arguments{
\item{dat}{matrix or data.frame, main input}

\item{lMod}{linear model, only used to extract coefficients offset & slope}

\item{level}{(numeric) alpha threshold for linear model}
}
\value{
This function returns a logical vector for each value in 2nd col of 'dat' if INSIDE confid interval
}
\description{
This function allows to compare 'dat' to confindence interval of linare model 'lMod' (eg from lm())
}
\examples{
set.seed(2016); dat1 <- matrix(c(runif(200)+rep(1:10,20)),ncol=10)
}
\seealso{
\code{\link{searchLinesAtGivenSlope}}
}
