% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorAccording2.R
\name{.chooseGrpCol}
\alias{.chooseGrpCol}
\title{Automatic choice of colors}
\usage{
.chooseGrpCol(
  nGrp,
  paired = FALSE,
  alph = 0.2,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{nGrp}{(numeric vector) main input}

\item{paired}{(logical)}

\item{alph}{(numeric vector)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced#'}
}
\value{
This function returns a character vector with color codes
}
\description{
This function allows to do automatic choice of colors: if single-> grey, if few -> RColorBrewer, if many : gradient green -> grey/red
}
\examples{
.chooseGrpCol(4)	
}
\seealso{
\code{\link[grDevices]{rgb}}; \code{\link{colorAccording2}}
}
