% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutArrayInCluLike.R
\name{.combineListAnnot}
\alias{.combineListAnnot}
\title{Combine annotation information from list of matrixes}
\usage{
.combineListAnnot(
  lst,
  useCol = 1:2,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{lst}{(list) main input}

\item{useCol}{(numeric vector) which columns should be used}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a single matrix of combined (non-redundant) info
}
\description{
This function allows to combine information (annotation) from list of matrixes (ie replace when NA), using always the columns specified in 'useCol' (numeric)
}
\examples{
.datSlope(c(3:6))
}
\seealso{
used in \code{\link{cutArrayInCluLike}}
}
