% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCloseMatch.R
\name{.compareByPPM}
\alias{.compareByPPM}
\title{Compare by PPM}
\usage{
.compareByPPM(dat, limit, distVal = FALSE)
}
\arguments{
\item{dat}{list of 2 numerical vectors}

\item{limit}{(numeric, length=1) threshold value for retaining values, used with distace-type specified in argument 'compTy'}

\item{distVal}{(logical) to toggle outpout as matrix of numeric (distance values above 'limit', others NA) or matrix of logical}
}
\value{
This function returns a list with close matches of 'x' to given 'y', the numeric value dependes on 'sortMatch' (if FASLE then always value of 'y' otherwise of longest of x&y)
}
\description{
This function allows to compare by ppm
}
\examples{
cc <- list(aa=11:14, bb=c(13.1,11.5,14.3,20:21))
.compareByPPM(cc, 1) 
}
\seealso{
\code{\link{findCloseMatch}},  \code{\link{checkSimValueInSer}}, and also \code{.compareByDiff}, for convient output \code{\link{countCloseToLimits}}
}
