% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowNormalize.R
\name{.complCols}
\alias{.complCols}
\title{Search (complementing) columns for best coverage of non-NA data for rowNormalization (main)}
\usage{
.complCols(x, dat, nCombin)
}
\arguments{
\item{x}{(integer, length=1) column number for with other columns to combine & give (some) complete non-NA lines are seeked}

\item{dat}{(matrix) .. init data, smay be parse matrix with numerous NA}

\item{nCombin}{(integer) .. number of columns used to make complete subset}
}
\value{
This function returns a matrix of column-indexes complementing (nCombin rows)
}
\description{
This function was designed to complete the selection of columns of sparse matrix 'dat' with sets of 'nCombin' columns at complete 'coverage'
Context : In sparse matrix 'dat' search subsets of columns with some rows as complete (no NA).
}
\examples{
.complCols(3, dat=matrix(c(NA,12:17,NA,19),ncol=3), nCombin=3)
}
\seealso{
\code{\link{rowNormalize}}
}
