% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equLenNumber.R
\name{.composeCallName}
\alias{.composeCallName}
\title{Compose sequence of (function-)calls}
\usage{
.composeCallName(newNa, add2Head = "", add2Tail = " : ", callFrom = NULL)
}
\arguments{
\item{newNa}{(character vector) main input}

\item{add2Head}{(character)}

\item{add2Tail}{(character)}

\item{callFrom}{(character) may also contain multiple separate names (ie length >1), will be concatenated using ' -> '}
}
\value{
character vector (history of who called whom)
}
\description{
This function was designed for tracing the hierarchy of function-calls.
It allows to remove any tailing space or ': ' from 'callFrom' (character vector) and return with added 'newNa' (+ 'add2Tail')
}
\examples{
.composeCallName("newFunction", callFrom="initFunction")
}
\seealso{
\code{\link[base]{paste}}
}
