% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVectLength.R
\name{.convertMatrToNum}
\alias{.convertMatrToNum}
\title{Convert numeric matrix to numeric}
\usage{
.convertMatrToNum(matr, useCol = NULL)
}
\arguments{
\item{matr}{(matrix) main input}

\item{useCol}{(integer) design the comumns to be used}
}
\value{
numeric vector
}
\description{
Take matrix and return vector
}
\examples{
.convertMatrToNum(matrix(1:6, ncol=2))
}
\seealso{
\code{\link[base]{matrix}}
}
