% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioAllComb.R
\name{.corDuplItemsByIncrem}
\alias{.corDuplItemsByIncrem}
\title{Avoid duplicating items between 'curNa' and 'newNa' by incrementing digits after 'extPref' (in newNa)}
\usage{
.corDuplItemsByIncrem(newNa, curNa, extPref = "_s")
}
\arguments{
\item{newNa}{(character) main input 1}

\item{curNa}{(character) main input 2}

\item{extPref}{(character) extension}
}
\value{
This function returns the corrected input vector \code{newNa}
}
\description{
This function aims to avoid duplicating items between 'curNa' and 'newNa' by incrementing digits after 'extPref' (in newNa)
}
\examples{
.corDuplItemsByIncrem(letters[1:6], letters[8:4])
}
\seealso{
\code{\link[base]{duplicated}}
}
