% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteC.R
\name{.cutStr}
\alias{.cutStr}
\title{Cut string to get all variants from given start with min and max length}
\usage{
.cutStr(txt, startFr = 1, minLe = 1, maxLe = NULL, reverse = TRUE)
}
\arguments{
\item{txt}{(character) main input, may be length >1}

\item{startFr}{(interger) where to start}

\item{minLe}{(interger) minimum length of output}

\item{maxLe}{(interger) maximum length of output}

\item{reverse}{(logical) return longest text-fragments at beginning of vector}
}
\value{
This function returns a character vector
}
\description{
This function allows truncating character vector to all variants from given start, with min and optonal max length
Used to evaluate argument calls without giving full length of argument
}
\examples{
.cutStr("abcdefg", minLe=2)
}
\seealso{
used in \code{\link{pasteC}}; \code{\link[base]{substr}}
}
