% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjBy2ptReg.R
\name{.datSlope}
\alias{.datSlope}
\title{Model linear regression and optional plot}
\usage{
.datSlope(
  dat,
  typeOfPlot = "sort",
  toNinX = FALSE,
  plotData = FALSE,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{dat}{(vector or matrix) main input}

\item{typeOfPlot}{(character)}

\item{toNinX}{(logical)}

\item{plotData}{(logical)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) display additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
numeric vector with intercept and slope, optional plot
}
\description{
This function allows to model a linear regression and optionally to plot the results
}
\examples{
.datSlope(c(3:6))
}
\seealso{
\code{\link[base]{append}}; \code{\link{lrbind}}
}
