% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstOfRepLines.R
\name{.extrNAneighb}
\alias{.extrNAneighb}
\title{Extract NA-neighbour values}
\usage{
.extrNAneighb(x, grp)
}
\arguments{
\item{x}{initial matrix to treat}

\item{grp}{(factor) grouing of replicates}
}
\value{
snumeric vector
}
\description{
This function allows extracting NA-neighbour value
}
\examples{
.extrNAneighb(c(11:14,NA), rep(1,5))
}
\seealso{
\code{\link[base]{unique}}, \code{\link{nonAmbiguousNum}}, faster than \code{\link{firstOfRepeated}} which gives more detail in output (lines/elements/indexes of omitted)
}
