% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nNonNumChar.R
\name{.extrNumHeadingCap}
\alias{.extrNumHeadingCap}
\title{Extract number(s) before capital character}
\usage{
.extrNumHeadingCap(x)
}
\arguments{
\item{x}{character vector to be treated}
}
\value{
This function returns a numeric vector
}
\description{
This function aims to extract number(s) before capital character
}
\examples{
.extrNumHeadingCap(" 1B ")
}
\seealso{
\code{\link[base]{grep}}, \code{\link[base]{nchar}}
}
