% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nNonNumChar.R
\name{.extrNumHeadingSepChar}
\alias{.extrNumHeadingSepChar}
\title{Extract numbers before separator followed by alphabetic character}
\usage{
.extrNumHeadingSepChar(x, sep = "_")
}
\arguments{
\item{x}{character vector to be treated}

\item{sep}{(character) separator}
}
\value{
This function returns a numeric vector
}
\description{
This function aims to extract number(s) before separator followed by alphabetic character (return named numeric vector, NAs when no numeric part found)
}
\examples{
.extrNumHeadingSepChar(" 1B ")
}
\seealso{
\code{\link[base]{nchar}}
}
