% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtSizeUniq.R
\name{.filtSize}
\alias{.filtSize}
\title{Filter for size}
\usage{
.filtSize(x, minSize = 5, maxSize = 36)
}
\arguments{
\item{x}{main inpuy}

\item{minSize}{(integer) minimum number of characters, if \code{NULL} set to 0}

\item{maxSize}{(integer) maximum number of characters}
}
\value{
list of filtered input
}
\description{
This function aims to filter for size
}
\examples{
aa <- 1:10
}
\seealso{
\code{\link{filtSizeUniq}}; \code{\link{correctToUnique}}, \code{\link[base]{unique}}, \code{\link[base]{duplicated}}
}
