% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialDist.R
\name{.findBorderOverlaps}
\alias{.findBorderOverlaps}
\title{Find overlap instances among range of values in lines}
\usage{
.findBorderOverlaps(x, rmRedund = FALSE, callFrom = NULL)
}
\arguments{
\item{x}{(matrix of numeric values or all-numeric data.frame) main input}

\item{rmRedund}{(logical) report overlaps only in 1st instance (will show up twice otherwise)}

\item{callFrom}{(character) allow easier tracking of message(s) produced}
}
\value{
This function returns a matrix with line for each overlap found, cols 'refLi' (line no), 'targLi' (line no), 'targCol' (col no)
}
\description{
This function aims to find overlap instances among range of values in lines of 'x' (typically give just min & max)
}
\examples{
aa <- 11:15
}
\seealso{
\code{\link[base]{nchar}}
}
