% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStrictOrder.R
\name{.firstMin}
\alias{.firstMin}
\title{Get first minimum}
\usage{
.firstMin(x, positionOnly = FALSE)
}
\arguments{
\item{x}{(numeric vector) main input}

\item{positionOnly}{(logical)}
}
\value{
numeric vector
}
\description{
This function allows to find the first minimum of a numeric vector
}
\examples{
.firstMin(c(4,3:6))
}
\seealso{
\code{\link[base]{which.min}}
}
