% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeAsListOfRepl.R
\name{.fuse2ArrBy2ndDim}
\alias{.fuse2ArrBy2ndDim}
\title{fuse 2 instances of 3dim arr as mult cols in 3dim array}
\usage{
.fuse2ArrBy2ndDim(arr1, arr2, silent = FALSE, debug = FALSE, callFrom = NULL)
}
\arguments{
\item{arr1}{(array)}

\item{arr2}{(array)}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This functuin returns a numeric vector with  numer of non-numeric characters (ie not '.' or 0-9))
}
\description{
This function allows fusing 2 instances of 3dim arr as mult cols in 3dim array (ie fuse along 2nd dim, increase cols)
}
\examples{
aa <- 11:15
}
\seealso{
\code{\link[base]{array}}
}
