% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMAList.R
\name{.getAmean}
\alias{.getAmean}
\title{Get A value for each group of replicates}
\usage{
.getAmean(dat, grp)
}
\arguments{
\item{dat}{(matrix or data.frame) main input}

\item{grp}{(factor) grouping of replicates}
}
\value{
This function returns a numeric vector
}
\description{
This function calculates the 'A' value (ie group mean) for each group of replicates (eg for MA-plot)
}
\examples{
.getAmean(matrix(11:18, ncol=4), gl(2,2))
}
\seealso{
\code{\link{makeMAList}}
}
