% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMAList.R
\name{.getAmean2}
\alias{.getAmean2}
\title{Get A value for each group of replicates based on comp}
\usage{
.getAmean2(dat, comp)
}
\arguments{
\item{dat}{(matrix or data.frame) main input}

\item{comp}{(matrix) tells which groups to use/compare, assuming that dat are already group-means)}
}
\value{
This function returns a numeric vector
}
\description{
This function calculates the 'A' value (ie group mean) for each group of replicates (eg for MA-plot)
\code{comp} is matrix telling which groups to use/compare, assuming that dat are already group-means)
}
\examples{
.getAmean(matrix(11:18, ncol=4), gl(2,2))
}
\seealso{
\code{\link{makeMAList}}
}
