% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTree.R
\name{.growTree}
\alias{.growTree}
\title{Grow tree}
\usage{
.growTree(tm, setX, addToObj = NULL)
}
\arguments{
\item{tm}{(list) main input, $disDat .. matrix with integer start & end sites for fragments; $lo (logical) which fragments may be grown; $start (integer) index for which line of $disDat to start; $it numeric version of $lo; $preN for previous tree objects towards root; $iter for iterator (starting at 1))}

\item{setX}{.. data.tree object (main obj from root)}

\item{addToObj}{.. data.tree object (branch on which to add new branches/nodes)}
}
\value{
list
}
\description{
This function allows growing tree-like structures (data.tree objects)
}
\examples{
.datSlope(c(3:6))
}
\seealso{
\code{\link{buildTree}}
}
