% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameColumns.R
\name{.keepFiniteCol}
\alias{.keepFiniteCol}
\title{Remove all columns where all data are not finite}
\usage{
.keepFiniteCol(
  dat,
  msgStart = NULL,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{dat}{(matrix or data.frame) main input}

\item{msgStart}{(character)}

\item{silent}{(logical) suppres messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allows easier tracking of messages produced}
}
\value{
This function returns a corrected matrix or data.frame
}
\description{
This function aims to remove all columns where all data are not finite
}
\examples{
ma1 <- matrix(c(1:5, Inf), ncol=2)
.keepFiniteCol(ma1)

}
\seealso{
\code{\link{renameColumns}}; \code{\link[base]{is.finite}}
}
