% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convMatr2df.R
\name{.mayBeNum}
\alias{.mayBeNum}
\title{Check if vector may be numeric content}
\usage{
.mayBeNum(x, pattern = NULL)
}
\arguments{
\item{x}{(numeric vector) main input}

\item{pattern}{(character)  custom pattern to check}
}
\value{
This functions returns a logical/boolean vector for each of the elements of 'x'
}
\description{
This function allows to checking if a given vector may be numeric content
}
\examples{
.mayBeNum(c(3:6))
}
\seealso{
\code{\link[base]{numeric}}; \code{\link{convMatr2df}}
}
