% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStrictOrder.R
\name{.medianSpecGrp}
\alias{.medianSpecGrp}
\title{Rescale respective to specific group}
\usage{
.medianSpecGrp(x, grpNum, grpVal, sumMeth = "median", callFrom = NULL)
}
\arguments{
\item{x}{(numeric vector) main input}

\item{grpNum}{(numeric)}

\item{grpVal}{(numeric)}

\item{sumMeth}{(character) method for summarizing}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
numeric vector
}
\description{
This function allows to rescale data 'x' so that specific group 'grpNum' gets normalized to predefined value 'grpVal'.
In normal case x will be multiplied by 'grpVal' and devided by value obtained from 'grpNum'.
If summary of 'grpNum-positions' or 'grpVal' is 0, then grpVal will be attained by subtraction of summary & adding grpVal
}
\examples{
.firstMin(c(4,3:6))
}
\seealso{
\code{\link[base]{which.min}}
}
