% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minDiff.R
\name{.minDif}
\alias{.minDif}
\title{find closest neighbour to numeric vector}
\usage{
.minDif(z, initOrder = TRUE, rat = TRUE)
}
\arguments{
\item{z}{(numeric) vector to search minimum difference}

\item{initOrder}{(logical) return matrix so that 'x' matches exactely 2nd col of output}

\item{rat}{(logical) express result as ratio}
}
\value{
This function returns a matrix with index,value,dif,best
}
\description{
This function aims to find closest neighbour to numeric vector
}
\examples{
.minDif(c(11:15,17))
}
\seealso{
\code{\link[stats]{dist}}
}
