% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusePairs.R
\name{.neigbDis}
\alias{.neigbDis}
\title{Distances beteenw sorted points of 2-columns}
\usage{
.neigbDis(x, asSum = TRUE)
}
\arguments{
\item{x}{(matrix or data.frame, min 2 columns) main input}

\item{asSum}{(logical) if \code{TRUE} (default) the sum of all distances will be returned, otherwise the individual distances}
}
\value{
This function returns a numeric vector with distances
}
\description{
This function returns distances beteenw sorted points of 2-column matrix 'x'
}
\examples{
daPa <- matrix(c(1:5,8,2:6,9), ncol=2)
.neigbDis(daPa)
}
