% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presenceFilt.R
\name{.offCenter}
\alias{.offCenter}
\title{Return position of 'di' (numeric vector) which is most excentric (distant to 0), starts with NAs as most excentric}
\usage{
.offCenter(di, nMost = 1)
}
\arguments{
\item{di}{(numeric) main input}

\item{nMost}{(integer)}
}
\value{
This function returns a integer/numeric vector (indicating index)
}
\description{
This function aims to return position of 'di' (numeric vector) which is most excentric (distant to 0), starts with NAs as most excentric
It is used for identifying/removing (potential) outliers. 
Note : this fx doesn't consider reference distrubutions, even with "perfect data" 'nMost' points will ba tagged !
}
\examples{
.offCenter(11:14)
}
\seealso{
use in \code{\link{presenceFilt}};  \code{\link[base]{diff}}
}
