% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstOfRepLines.R
\name{.pasteCols}
\alias{.pasteCols}
\title{Paste-concatenate all columns of matrix}
\usage{
.pasteCols(mat, sep = "")
}
\arguments{
\item{mat}{inital matrix}

\item{sep}{(character) separator}
}
\value{
simplified/non-redundant vector/matrix (ie fewer lines for matrix), or respective index
}
\description{
This function allows paste columns
}
\examples{
.pasteCols(matrix(11:16,ncol=2), sep="_")
}
\seealso{
\code{\link[base]{unique}}, \code{\link{nonAmbiguousNum}}, faster than \code{\link{firstOfRepeated}} which gives more detail in output (lines/elements/indexes of omitted)
}
