% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVectLength.R
\name{.plusLowerCaps}
\alias{.plusLowerCaps}
\title{Add lower caps to character vector}
\usage{
.plusLowerCaps(x)
}
\arguments{
\item{x}{(character) main input}
}
\value{
This function returns a elongated character vector
}
\description{
This function allows adding all content as lower caps to/of character vector
}
\examples{
.plusLowerCaps(c("Abc","BCD"))
}
\seealso{
\code{\link[base]{chartr}}
}
