% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialDist.R
\name{.raiseColLowest}
\alias{.raiseColLowest}
\title{Raise all values close to lowest value}
\usage{
.raiseColLowest(
  mat,
  raiseTo = NULL,
  minFa = 0.1,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{mat}{(matrix of numeric values) main input}

\item{raiseTo}{(numeric)}

\item{minFa}{(numeric) minimum factor}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) display additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of message(s) produced}
}
\value{
This function returns a numeric vector with  numer of non-numeric characters (ie not '.' or 0-9))
}
\description{
This function aims to raise all values close to lowest value to end up as at value of 'raiseTo'.
This is done independently for each col of mat.
This function sets all data to common raiseTo (which is min among all cols)
}
\examples{
aa <- 11:15
}
\seealso{
\code{\link[base]{nchar}}
}
