% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioAllComb.R
\name{.removeCol}
\alias{.removeCol}
\title{Remove columns indicated by col-number}
\usage{
.removeCol(matr, rmCol)
}
\arguments{
\item{matr}{(matrix or data.frame) main input}

\item{rmCol}{(integer) column index for removing}
}
\value{
This function returns an matrix or data.frame
}
\description{
This function aims to remove columns indicated by col-number
}
\examples{
aa <- matrix(1:6, ncol=3)
.removeCol(aa, 2)
}
\seealso{
\code{\link[stats]{dist}}
}
