% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVectLength.R
\name{.replSpecChar}
\alias{.replSpecChar}
\title{Replace Special Characters}
\usage{
.replSpecChar(x, findSp = c("\\\\(", "\\\\)", "\\\\$"), replBy = "_")
}
\arguments{
\item{x}{(character) main input}

\item{findSp}{(character) special characters to replace (may have to be given as protected)}

\item{replBy}{(character) replace by}
}
\value{
This function returns a corrceted/adjusted factor
}
\description{
This function allows replacing special characters
Note that (most) special characters must be presented with protection for \code{grep} and \code{sub}.
}
\examples{
.replSpecChar(c("jhjh(ab)","abc"))
}
\seealso{
\code{\link[base]{factor}}
}
