% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrSpcText.R
\name{.retain1stPart}
\alias{.retain1stPart}
\title{Trim character string: keep only text before 'sep'}
\usage{
.retain1stPart(chr, sep = " = ", offSet = 1)
}
\arguments{
\item{chr}{character vector to be treated}

\item{sep}{(character) saparator}

\item{offSet}{(integer) off-set}
}
\value{
This function returns a modified character vector
}
\description{
Trim character string: keep only text before 'sep' (length=1 !)
}
\examples{
.retain1stPart("abc = def")
}
\seealso{
\code{\link[base]{substr}}
}
