% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleXY.R
\name{.scale01}
\alias{.scale01}
\title{Scale between 0 and 1 (main)}
\usage{
.scale01(x)
}
\arguments{
\item{x}{numeric vector to be re-scalded}
}
\value{
This function returns a numeric vector of same length with re-scaled values
}
\description{
This function rescales between 0 and 1
}
\examples{
.scale01(11:15)
}
\seealso{
\code{\link{scaleXY}} , \code{\link[base]{scale}}
}
