% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleXY.R
\name{.scaleXY}
\alias{.scaleXY}
\title{Scale between min and max value (main)}
\usage{
.scaleXY(x, minim = 2, maxim = 3)
}
\arguments{
\item{x}{numeric vector to be re-scalded}

\item{minim}{(numeric) minimum value for resultant vactor}

\item{maxim}{(numeric) minimum value for resultant vactor}
}
\value{
This function returns a matrix of CV values
}
\description{
This function rescales between user-defined min and max values
}
\examples{
.scaleXY(11:15, min=1, max=100)
}
\seealso{
\code{\link{scaleXY}} , \code{\link[base]{scale}}
}
