% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteC.R
\name{.seqCutStr}
\alias{.seqCutStr}
\title{Cut string to get all variants from given start with min length, depreciated}
\usage{
.seqCutStr(txt, startFr = 1, minLe = 1, reverse = TRUE)
}
\arguments{
\item{txt}{(character) main input, may be length >1}

\item{startFr}{(interger) where to start}

\item{minLe}{(interger) minimum length of output}

\item{reverse}{(logical) return longest text-fragments at beginning of vector}
}
\value{
This function returns a character vector
}
\description{
This function is depreciated, please use \code{/cutStr} instead !
This function allows truncating character vector to all variants from given start, with min and optonal max length
Used to evaluate argument calls without giving full length of argument
}
\examples{
.seqCutStr("abcdefg", minLe=2)
}
\seealso{
\code{\link{pasteC}}; \code{\link[base]{substr}}
}
