% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioAllComb.R
\name{.stackArray}
\alias{.stackArray}
\title{Reorganize array by reducing dimension 'byDim'  (similar to stack() for data-frames)}
\usage{
.stackArray(arr, byDim = 3)
}
\arguments{
\item{arr}{(array) main input}

\item{byDim}{(integer)}
}
\value{
This function returns an array/matrix of 1 dimension less than 'arr', 1st dim has more lines (names as paste with '_')
}
\description{
This function aims to reorganize an array by reducing dimension 'byDim'  (similar to stack() for data-frames)
It returns an array/matrix of 1 dimension less than 'arr', 1st dim has more lines (names as paste with '_')
}
\examples{
(arr1 <- array(11:37, dim=c(3,3,3)))
.stackArray(arr1, 3)
}
\seealso{
\code{\link[stats]{dist}}
}
