% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimRedundText.R
\name{.trimFromEnd}
\alias{.trimFromEnd}
\title{Trim from end (Deprecated)}
\usage{
.trimFromEnd(x, ..., callFrom = NULL, debug = FALSE, silent = TRUE)
}
\arguments{
\item{x}{character vector to be treated}

\item{...}{more vectors to be treated}

\item{callFrom}{(character) allow easier tracking of messages produced}

\item{debug}{(logical) display additional messages for debugging}

\item{silent}{(logical) suppress messages}
}
\value{
This function returns a modified character vector
}
\description{
Deprecated Version - This function allows trimming/removing redundant text-fragments from end
}
\examples{
txt1 <- c("abcd_ccc","bcd_ccc","cde_ccc")
.trimRight(txt1)
}
\seealso{
\code{\link{trimRedundText}}; Inverse : Find/keep common text \code{\link{keepCommonText}};  you may also look for related functions in package \href{https://CRAN.R-project.org/package=stringr}{stringr}
}
