% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimRedundText.R
\name{.trimFromStart}
\alias{.trimFromStart}
\title{Trim from start (Deprecated)}
\usage{
.trimFromStart(
  x,
  ...,
  minNchar = 1,
  silent = TRUE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{x}{character vector to be treated}

\item{...}{more vectors to be treated}

\item{minNchar}{(integer) minumin number of characters that must remain}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) display additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a modified character vector
}
\description{
Deprecated Version - This function allows trimming/removing redundant text-fragments from start
}
\examples{
txt1 <- c("abcd_ccc","bcd_ccc","cde_ccc")
.trimLeft(txt1)  # replacement
}
\seealso{
\code{\link{trimRedundText}}; Inverse : Find/keep common text \code{\link{keepCommonText}};  you may also look for related functions in package \href{https://CRAN.R-project.org/package=stringr}{stringr}
}
