% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVectLength.R
\name{.vector2Matr}
\alias{.vector2Matr}
\title{Convert numeric vector to matrix}
\usage{
.vector2Matr(x, colNa = NULL, rowsKeep = TRUE)
}
\arguments{
\item{x}{(numeric or character) main input}

\item{colNa}{(integer) design the comumn-name to be given}

\item{rowsKeep}{(logical) is \code{TRUE} make matrix of 1 column, otherwise of 1 row}
}
\value{
matrix
}
\description{
Take (numeric) vector and return matrix, if 'colNa' given will be used as colname
}
\examples{
.vector2Matr(c(3:6))
}
\seealso{
\code{\link[base]{matrix}}
}
