% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLast2numericParts.R
\name{extractLast2numericParts}
\alias{extractLast2numericParts}
\title{Extract Last Two Numeric Parts From Character Vector}
\usage{
extractLast2numericParts(x, silent = FALSE, debug = FALSE, callFrom = NULL)
}
\arguments{
\item{x}{main character input}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a (numeric) matrix with 2 columns (eg from initial concatenated coordinates)
}
\description{
This function extracts last 2 (integer) numeric parts between punctuations out of character vector 'x'.
Runs faster than \code{gregexpr} .
Note: won't work correctly with decimals or exponential signs !! (such characters will be considered as punctuation, ie as separator)
}
\examples{
extractLast2numericParts(c("M01.1-4","M001/2.5","M_0001_03-16","zyx","012","a1.b2.3-7,2"))
}
\seealso{
\code{gregexpr} from \code{\link[base]{grep}}
}
