% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRepeated.R
\name{findRepeated}
\alias{findRepeated}
\title{Find Repeated Elements}
\usage{
findRepeated(
  x,
  nonRepeated = FALSE,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{x}{character vector}

\item{nonRepeated}{(logical) if \code{nonRepeated=TRUE} return list with elements \code{$rep} and \code{$nonrep}}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a (named) list of indexes for each of the repeated values, or NULL if all values unique
}
\description{
This function gets index of repeated items/values in vector 'x' (will be treated as character). 
It returns (named) list of indexes for each of the repeated values, or \code{NULL} if all values are unique.
This approach is similar but more basic compared to \code{\link{get1stOfRepeatedByCol}}.
}
\examples{
aa <- c(11:16,14:12,14); findRepeated(aa) 
}
\seealso{
similar approach but more basic than \code{\link{get1stOfRepeatedByCol}}
}
