% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrbind.R
\name{lrbind}
\alias{lrbind}
\title{rbind on lists}
\usage{
lrbind(lst, silent = FALSE, callFrom = NULL)
}
\arguments{
\item{lst}{(list) main input}

\item{silent}{(logical) suppress messages}

\item{callFrom}{(character) allow easier tracking of message produced}
}
\value{
matrix or data.frame
}
\description{
rbind-like function to append list-elements containing tables and return one long table. 
Accepts also list-entries with data.frames or vectors (of length of no of columns) as long as at least 1 list-entry is a matrix
}
\examples{
lst1 <- list(matrix(1:9,nc=3,dimnames=list(letters[1:3],c("AA","BB","CC"))),
  11:13,matrix(51:56,ncol=3))
lrbind(lst1)
}
\seealso{
\code{rbind} in \code{\link[base]{cbind}}
}
