% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naOmit.R
\name{naOmit}
\alias{naOmit}
\title{Fast na.omit}
\usage{
naOmit(x, silent = FALSE, debug = FALSE, callFrom = NULL)
}
\arguments{
\item{x}{(vector or matrix) data to check & remove \code{NA}s}

\item{silent}{(logical) suppress messages if \code{TRUE}}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allows easier tracking of messages produced}
}
\value{
This function returns a vector without NAs (matrix input will be transformed to vector). Returns NULL if input consists only of NAs.
}
\description{
This function removes NAs from input vector, in contrast to \code{na.omit} this function has no slot for removed elements.
}
\details{
Resulting objects from \code{naOmit} are smaller in size and subsequent code execution (on large vectors) may be faster (in particular if many NAs get encountered).
Note : This function behaves differently to \code{na.omit} with input other than plain vectors. Will not work with data.frames !
}
\examples{
aA <- c(11:13,NA,10,NA); 
naOmit(aA) 
}
\seealso{
\code{\link[stats]{na.fail}}, \code{na.omit}
}
