% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nFragments.R
\name{nFragments}
\alias{nFragments}
\title{Number Of Fragments After Cut At Specific Character(s) Within Size-range}
\usage{
nFragments(
  protSeq,
  cutAt,
  sizeRa,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{protSeq}{(character)  text to be cut}

\item{cutAt}{(character) position to cut}

\item{sizeRa}{(numeric,length=2) min and max size to consider}

\item{silent}{(logical) suppress messages if \code{TRUE}}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allows easier tracking of messages produced}
}
\value{
This function returns a numeric vector with number of fragments for each entry 'protSeq' (names are 'protSeq')
}
\description{
This function determines number of fragments /entry within range of 'sizeRa' (numeric,length=2) when cutting after 'cutAt'
}
\examples{
tmp <- "MSVSREDSCELDLVYVTERIIAVSFPSTANEENFRSNLREVAQMLKSKHGGNYLLFNLSERRPDITKLHAKVLEFGWPDLHTPALEKI"
nFragments(c(tmp,"ojioRij"),c("R","K"),c(4,31))

}
\seealso{
\code{\link{cutAtMultSites}}, simple version \code{{nFragments0}} (no size-range)
}
