% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partUnlist.R
\name{partUnlist}
\alias{partUnlist}
\title{Partial unlist of lists of lists}
\usage{
partUnlist(lst)
}
\arguments{
\item{lst}{list to be partailly unlisted}
}
\value{
list with partially reduced nested structure
}
\description{
\code{partUnlist} does  partial unlist for treating list of lists : New (returned) list has one level less of hierarchy 
(Highest level list will be appended). In case of conflicting (non-null) listnames a prefix will be added. 
Behaviour different to \code{\link[base]{unlist}} when unlisting list of matrixes.
}
\examples{
partUnlist(list(list(a=11:12,b=21:24),list(c=101:101,d=201:204)))
partUnlist(list(c=1:3,L2=matrix(1:4,ncol=2),list(L1=11:12,L2=matrix(21:26,ncol=2))))
unlist(list(c=1:3,L2=matrix(1:4,ncol=2),list(L1=11:12,L2=matrix(21:26,ncol=2))),rec=FALSE)
}
\seealso{
\code{\link[base]{unlist}}, \code{\link{asSepList}}
}
