% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteC.R
\name{pasteC}
\alias{pasteC}
\title{Advanced paste-collapse}
\usage{
pasteC(..., collapse = ", ", lastCol = " and ", quoteC = "")
}
\arguments{
\item{...}{(character) main input to be collapsed}

\item{collapse}{(character,length=1) element to use for collapsing}

\item{lastCol}{(character) text to use before last item enumerated element}

\item{quoteC}{character to use for citing with quotations (default "")}
}
\value{
This function returns a character vector of truncated versions of intpup \code{txt}
}
\description{
This function is a variant of \code{\link[base]{paste}} for convenient use of paste-collapse and separation of last element to paste (via 'lastCol').
This function was mode for more human like enumeriating in output and messages.
If multiple arguments are given without names they will all be concatenated, if they contain names lazy evaluation for names will be tried
(with preference to longest match to argument names).
Note that some special characters (like backslash) may need to be protetected when used with 'collapse' or 'quoteC'.
Returns character vector of length 1 (everything pasted together)
}
\examples{
pasteC(1:4)
}
\seealso{
\code{\link[base]{paste}} for basic paste
}
\keyword{character}
