% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwSeparatorList.R
\name{pwSeparatorList}
\alias{pwSeparatorList}
\title{Suggestions Of Collections Of Classical Separators For Pairwise Combinations}
\usage{
pwSeparatorList(
  type = "split1",
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{type}{(character) allows selecting of different types of collections of separators}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a charcter vector with proposed separators for later choosing the most suited
}
\description{
This function provides collecions of typical separators used for combined names in pair-wise testing for later choice of the most suited.
}
\details{
Depending if one rather looks for separators for cutting existing names of for suggesting separators for fresh concatenation, two differentsets of collections are available via the argument \code{type}.

Using \code{type="split"} will default to \code{type="split1"}, similarly \code{type="combine"} will be used as \code{type="combine1"}.
  
The option \code{type="split1"} has a default collection of shorter/compact separators for splitting
The option \code{type="split2"} allows more extensive search for 38 separators (ie combinations with ' ','-' and '_'), but extensive checking may cost CPU time.
The option \code{type="split3"} has a shorter collection of 12 basic commonly used separators.
The option \code{type="split4"} to  \code{type="split9"} default to 'split1'

The option \code{type="combine1"} does not suggest '--' (neither '__') since if '-' may not be chosen  ins the first place, using '--' as separator next to '-' occuring inside gourp-names may not enhance visibility.
The option \code{type="combine2"} contains a collection of sperators using 'vs' or 'versus'
The option \code{type="combine3"} is designed for explicite terms (starts with ' - versus - ', for suggesting compact terms rather use \code{type="combine1"}
The option \code{type="combine4"} to  \code{type="combine9"} default to 'combine1'
}
\examples{
pwSeparatorList(type="split1")
}
\seealso{
(for running multiple pair-wise test) \code{\link{moderTestXgrp}}, \code{\link{convPairwiseSetup}}, \code{\link{presenceFilt}}, \code{\link{getPWseparator}}, \code{\link{indexGroupsFromPW}}, 
  (used underneith:) \code{\link[base]{grep}}, \code{\link[base]{strsplit}}
}
