% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexGroupsFromPW.R
\name{replacePWseparator}
\alias{replacePWseparator}
\title{Replace Separator In Vector Of Pairwise Group-Names  
  
This function allows identifying and substituting a separator used in a character vector concatenated of pairwise groups.}
\usage{
replacePWseparator(
  compNames,
  newSep,
  potSep = c("-", "---", "_", "___", ".", "-vs-", "_vs_", "--vs--", "__vs__", " ", "  "),
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{compNames}{(character or integer) names of pairwise combined group-names}

\item{newSep}{(character of length=1) new separator between group-names for pair of names; if \code{NULL} a suitable separator (among '-','--','_','__','.',' ' and '  ') will be used}

\item{potSep}{(character) potential separators to be checked if occuring in \code{useComp}; the first fitting will used}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a character vector with group-wise substituted separators (or \code{NULL} if conflicts appear)
}
\description{
Replace Separator In Vector Of Pairwise Group-Names  
  
This function allows identifying and substituting a separator used in a character vector concatenated of pairwise groups.
}
\examples{
replacePWseparator(compNames=c("B-C","B-D","C-D"), newSep="-vs-") 
}
\seealso{
\code{\link{indexGroupsFromPW}}, \code{\link{getPWseparator}}, \code{\link{.getPWseparator}}
}
