% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysDate.R
\name{sysDate}
\alias{sysDate}
\title{System-data (compressed format)}
\usage{
sysDate(style = "wr")
}
\arguments{
\item{style}{(character) choose style (if \code{wr} the month will be abbreviated to fisrt 3 letters and last two digits of year will be shown}
}
\value{
character vector with formatted date
}
\description{
This function returns current date (based on Sys.Date) as day/month/year, where month is abbrev to 3 leters and only the mast two dgits of the year are shown.
}
\examples{
sysDate() 

}
\seealso{
\code{\link[base]{date}}, \code{Sys.Date} and \code{\link[base]{Sys.time}},
}
