% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top3mean.R
\name{top3mean}
\alias{top3mean}
\title{Mean Of 3 Highest Values}
\usage{
top3mean(x, silent = FALSE, debug = FALSE, callFrom = NULL)
}
\arguments{
\item{x}{(numeric vector) main input}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) additional messages for debugging}

\item{callFrom}{(character) allows easier tracking of messages produced}
}
\value{
This function returns a vector with single numeric value for mean of top3
  ( returns \code{NULL} with invalid input or if all input is \code{NA})
}
\description{
This function returns the mean of the top3 highest values.
}
\details{
Generally, \code{NA} will be excluded, if all values are \code{NA} this finction will return \code{NULL} ;
thus, in case of (entirely) unsuitable data (non-numeric ...)  \code{NULL} will be returned.
If data has subgroups you may try a tapply -way.
}
\examples{
x1 <- c(15:11,NA,16)
top3mean(x1)
}
\seealso{
\code{\link[base]{mean}}
}
