% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateVect.R
\name{truncateVect}
\alias{truncateVect}
\title{Truncating Text Vector To Specific Length}
\usage{
truncateVect(
  x,
  nMax = 20,
  sep = NULL,
  appendChar = "...",
  startFrom = 1,
  silent = FALSE,
  debug = FALSE,
  callFrom = NULL
)
}
\arguments{
\item{x}{(character or matrix of text) main input, text to be truncated}

\item{nMax}{(integer, length=1) max length in number of characters to be returned}

\item{sep}{(character or \code{NULL}) optional separator to be considered : Truncation will happen before the separator; set to  \code{NULL} for ignoring}

\item{appendChar}{(character, length=1) optional set of characters to add at end when trucation has happened (default '...'); set to  \code{NULL} or  \code{""} for not appending anything}

\item{startFrom}{(integer, length=1) designes the first character to be extracted (see also \code{\link[base]{substr}} )}

\item{silent}{(logical) suppress messages}

\item{debug}{(logical) display additional messages for debugging}

\item{callFrom}{(character) allow easier tracking of messages produced}
}
\value{
This function returns a data.frame or matrix (if no annotation added) with values (and annotation) conform to fiter criteria
}
\description{
This function allows truncating a text vector (or matrix of text) at nMax number of characters with optional adding of truncation/ending sign and trimming after last separator.
}
\details{
convenient access to results produced using the functions \code{\link[base]{substr}} or \code{\link{moderTest2grp}}.
The user can define the threshold and which type of multiple testing correction should be used
 (as long as the multiple testing correction method cited was actually performed as part of testing).


This function builds on \code{\link[base]{substr}} but offers additional options for trimming at right side of text:
In contrast to \code{\link[base]{substr}} where the absolute position number has to be given, this function takes 
the max desired length of the resulting character vector (argument \code{nMax}).
The argument \code{appendChar} allows adding extra characters (default '...') indicating that trimmming has been performed.
When doing so, the text will be trimmed further inside to ensure the final length of the resulting text, as defined by argument \code{nMax}.

Furthermore, using the argument \code{sep} it is possible to respect entities between separators and only cut at/before such separators
(the resulting text vecor may thus be shorter than specified globally with argument \code{nMax}).
}
\examples{
txt <- c("abc", "abcdefgh", "abc/def/gh")
truncateVect(txt, nMax=7, appendChar="..")
truncateVect(txt, nMax=7, sep="/", appendChar="..")

}
\seealso{
\code{\link[base]{substr}}
}
