% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBeforFileExtension.R
\name{.checkFileNameExtensions}
\alias{.checkFileNameExtensions}
\title{checkFileNameExtensions 
Function for checking file-names.}
\usage{
.checkFileNameExtensions(fileNa, ext)
}
\arguments{
\item{fileNa}{(character) file name to be checked}

\item{ext}{(character) file extension}
}
\value{
modified character vector
}
\description{
checkFileNameExtensions 
Function for checking file-names.
}
\examples{
.checkFileNameExtensions("testFile.txt","txt")
}
