% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCols.R
\name{.summarizeCols}
\alias{.summarizeCols}
\title{Summarize columns of matrix (or data.frame) 'x' using apply (main)}
\usage{
.summarizeCols(
  x,
  me = c("median", "medianComplete", "mean", "meanComplete", "aver", "average", "min",
    "max", "maxOfRef", "minOfRef", "maxAbsOfRef", "lastLi", "last", "firstComplete",
    "first", "firstLi", "summary"),
  vectAs1row = TRUE
)
}
\arguments{
\item{x}{data.frame matrix of data to be summarized by comlumn}

\item{me}{(character, length=1) summarization method (eg 'maxLast','minLast','maxLast','maxAbsLast', 'minLast', 'medianComplete' or 'meanComplete')}

\item{vectAs1row}{(logical) if TRUE will interprete non-matrix 'x' as matrix with 1 row (correct effect of automatic conversion when extracting 1 line)}
}
\value{
vector with summary for each column
}
\description{
This function summarizes columns of matrix (or data.frame) 'x' using apply
Note, it cannot handle character entries !  (all results will be NA)
}
\examples{
t1 <- matrix(round(runif(30,1,9)), nc=3); rownames(t1) <- letters[c(1:5,3:4,6:4)]  
}
\seealso{
\code{summarizeCols}
}
