% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equLenNumber.R
\name{equLenNumber}
\alias{equLenNumber}
\title{Equal character-length number}
\usage{
equLenNumber(x, silent = FALSE, callFrom = NULL)
}
\arguments{
\item{x}{(caracter) input vector}

\item{silent}{(logical) suppress messages}

\item{callFrom}{(character) allows easier tracking of message(s) produced}
}
\value{
character vector formated as equal number of characters per value
}
\description{
\code{equLenNumber} convert numeric entry 'x' to text, with all elements getting the same number of characters (ie  by adding preceeding or tailing 0s, if needed).
So far, the function cannot handle scientific annotations.
}
\examples{
equLenNumber(c(12,-3,321))
equLenNumber(c(12,-3.3,321))
}
\seealso{
\code{\link[base]{sprintf}}
}
