% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitDataUrl.R
\name{gitDataUrl}
\alias{gitDataUrl}
\title{Convert ulr-name for reading in raw-mode}
\usage{
gitDataUrl(urlName, replTxt = NULL, silent = FALSE, callFrom = NULL)
}
\arguments{
\item{urlName}{(charachter) main url-address}

\item{replTxt}{(NULL or matrix) adjust/ custom-modify search- and replacement items; should be matrix with 2 columns, 
the 1st colimn entries will be used as 'search-for' and the 2nd as 'replace by' fro each row.}

\item{silent}{(logical) suppress messages}

\item{callFrom}{(character) allow easier tracking of message(s) produced}
}
\value{
corrected urlName
}
\description{
This functions converts a given urlName so that from data from git-hub can be read correctly that tabular data.
Thus, this will remove '/blob/' and change starting characters to 'raw.githubusercontent.com'
}
\examples{
url1 <- paste0("https://github.com/bigbio/proteomics-metadata-standard/blob/",
  "master/annotated-projects/PXD001819/PXD001819.sdrf.tsv") 
gitDataUrl(url1)



}
\seealso{
\code{\link[base]{sub}};
}
