% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlSpecCharConv.R
\name{htmlSpecCharConv}
\alias{htmlSpecCharConv}
\title{Html special character conversion}
\usage{
htmlSpecCharConv(txt)
}
\arguments{
\item{txt}{character vector inclusing special characters}
}
\value{
corrected character vector adopted to html display
}
\description{
Converts 'txt' so that special characters (like 'beta','micro','square' etc) will be displayed correctly whe used for display in html (eg at mouse-over).
Note : uses package \href{https://CRAN.R-project.org/package=stringi}{stringi}
}
\examples{
(x <- stringi::stri_unescape_unicode("\\\\u00b5\\\\u003d\\\\u0061\\\\u0062"))
htmlSpecCharConv(x)
}
\seealso{
tables on \url{http://www.htmlhelp.com/reference/html40/entities/latin1.html},  \url{http://www.degraeve.com/reference/specialcharacters.php},  \url{http://www.ascii.cl/htmlcodes.htm}
}
