% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nFragments.R
\name{nFragments0}
\alias{nFragments0}
\title{Number of fragments after cut at specific character(s)}
\usage{
nFragments0(protSeq, cutAt)
}
\arguments{
\item{protSeq}{(character) text to be cut}

\item{cutAt}{(integer) position to cut}
}
\value{
numeric vector with number of fragments for each entry 'protSeq' (names are 'protSeq')
}
\description{
\code{nFragments0} tells the number of fragments/entry when cutting after 'cutAt'
}
\examples{
tmp <- "MSVSRTMEDSCELDLVYVTERIIAVSFPSTANEENFRSNLREVAQMLKSKHGGNYLLFNLSERRPDITKLHAKVLEFGWPDLHTPALEKI" 
nFragments0(c(tmp,"ojioRij"),c("R","K"))
}
\seealso{
more elaborate \code{{nFragments}}; \code{\link{cutAtMultSites}}
}
