% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowCVs.R
\name{rowCVs}
\alias{rowCVs}
\title{rowCVs}
\usage{
rowCVs(dat, autoconvert = NULL)
}
\arguments{
\item{dat}{(numeric) matix}

\item{autoconvert}{(NULL or character) allows converting simple vectors in matrix of 1 row (autoconvert="row")}
}
\value{
(numeric) vector with CVs for each row of 'dat'
}
\description{
\code{rowCVs} returns CV for values in each row (using speed optimized standard deviation).
Note : NaN values get replaced by NA.
}
\examples{
set.seed(2016); dat1 <- matrix(c(runif(200)+rep(1:10,20)),ncol=10)
head(rowCVs(dat1))
}
\seealso{
\code{\link[base]{colSums}}, \code{\link{rowGrpCV}}, \code{\link{rowSds}}
}
