% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triCoord.R
\name{triCoord}
\alias{triCoord}
\title{Pairwise x,y combinations}
\usage{
triCoord(n, side = "upper")
}
\arguments{
\item{n}{(integer) number of elements for making all pair-wise combinations}

\item{side}{(character) "upper" or "lower"}
}
\value{
2-column matrix wiyh indexes for all pairwise combnations of 1:n
}
\description{
\code{triCoord} gets pairwise combinations for 'n' elements; returns matrix with x & y coordinates to form all pairwise groups for 1:n elements
}
\examples{
triCoord(4)
}
\seealso{
\code{\link[base]{lower.tri}} or \code{upper.tri}, simpler version \code{\link{upperMaCoord}}
}
